requirements.txt
pyyaml==6.0.1
kubernetes==28.1.0
urllib3==1.26.18
